pyproject.toml
[tool.poetry]
name = "rezai"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
gunicorn = "^21.2.0"
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
alembic = "^1.11.1"
asyncpg = {version = "^0.28.0", extras = ["sa"]}
redis = {version = "^4.6.0", extras = ["hiredis"]}
aiofiles = "^23.1.0"
httptools = "^0.6.0"
loguru = "^0.7.0"
langchain-anthropic = "^0.1.19"
openai = "^1.35.13"
langchain-core = "^0.2.13"
langgraph = "^0.1.7"
twilio = "^9.2.3"
tavily-python = "^0.3.3"
langchain-community = "^0.2.7"
importlib_metadata = "^4.0.0"
setuptools = "^71.0.1"
streamlit = "^1.37.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~7"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.19.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"
setuptools = "^71.0.1"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.374"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["rezai",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "REZAI_ENVIRONMENT=pytest",
    "REZAI_DB_BASE=rezai_test",
]

[fastapi-template.options]
project_name = "rezai"
api_type = "rest"
enable_redis = "True"
enable_rmq = "None"
ci_type = "github"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "rezai"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "True"
orm = "sqlalchemy"
self_hosted_swagger = "True"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "True"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
